split server into functions
