+++ dp_creator_ii/app/analysis_panel.py
@@ -1,4 +1,4 @@
-from shiny import ui
+from shiny import ui, reactive
 
 
 def analysis_ui():
@@ -8,3 +8,10 @@ def analysis_ui():
         ui.input_action_button("go_to_results", "Download results"),
         value="analysis_panel",
     )
+
+
+def analysis_server(input, output, session):
+    @reactive.effect
+    @reactive.event(input.go_to_results)
+    def go_to_results():
+        ui.update_navs("top_level_nav", selected="results_panel")
